refactor(backend): Improve logging and exception handling
Added logic to throw exception when user tries to follow or unfollow himself.

+++ backend/src/main/java/com/github/ferigeek/sarv/service/FollowService.java
@@ -7,15 +7,16 @@ import com.github.ferigeek.sarv.exception.FollowException;
 import com.github.ferigeek.sarv.exception.UserNotFoundException;
 import com.github.ferigeek.sarv.repository.FollowRepository;
 import com.github.ferigeek.sarv.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.OffsetDateTime;
-import java.util.List;
 
 @Service
+@Slf4j
 public class FollowService {
 
     private final UserRepository userRepository;
@@ -29,7 +30,7 @@ public class FollowService {
 
     public Page<UserSummaryResponse> getFollowers(Long userId, Pageable pageable) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User not found with ID: <%d>".formatted(userId)));
+                .orElseThrow(() -> new UserNotFoundException("User not found with ID: %d".formatted(userId)));
         return followRepository.findByFollowed(user, pageable)
                 .map(Follow::getFollower)
                 .map(UserSummaryResponse::new);
@@ -37,7 +38,7 @@ public class FollowService {
 
     public Page<UserSummaryResponse> getFollowing(Long userId, Pageable pageable) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new UserNotFoundException("User not found with ID: <%d>".formatted(userId)));
+                .orElseThrow(() -> new UserNotFoundException("User not found with ID: %d".formatted(userId)));
         return followRepository.findByFollower(user, pageable)
                 .map(Follow::getFollowed)
                 .map(UserSummaryResponse::new);
@@ -46,35 +47,47 @@ public class FollowService {
     public void followUser(String username, Long userId) {
         User follower = userRepository.findByUsername(username)
                 .orElseThrow(() -> new UserNotFoundException(
-                        "Follower user not found with username: <%s>".formatted(username))
+                        "Follower user not found with username: %s".formatted(username))
                 );
         User followed = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(
-                        "Followed user not found with ID: <%d>".formatted(userId))
+                        "Followed user not found with ID: %d".formatted(userId))
                 );
+
+        if (follower.getId().equals(followed.getId())) {
+            throw new IllegalArgumentException("User can not follow himself ID: %d".formatted(userId));
+        }
 
         Follow follow = new Follow();
         follow.setFollower(follower);
         follow.setFollowed(followed);
         follow.setCreatedAt(OffsetDateTime.now());
         followRepository.save(follow);
+
+        log.info("User with ID={} followed user with ID={}", follower.getId(), followed.getId());
     }
 
     public void unfollowUser(String username, Long userId) {
         User follower = userRepository.findByUsername(username)
                 .orElseThrow(() -> new UserNotFoundException(
-                        "Follower user not found with username: <%s>".formatted(username))
+                        "Follower user not found with username: %s".formatted(username))
                 );
         User followed = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(
-                        "Followed user not found with ID: <%d>".formatted(userId))
+                        "Followed user not found with ID: %d".formatted(userId))
                 );
+
+        if (follower.getId().equals(followed.getId())) {
+            throw new IllegalArgumentException("User can not unfollow himself ID: %d".formatted(userId));
+        }
 
         Follow follow = followRepository.findByFollowerAndFollowed(follower, followed)
                 .orElseThrow(() -> new FollowException(
-                        "A follow from user with ID: <%d>, following user with ID: <%d>, doesn't exist"
+                        "A follow from user with ID: %d, following user with ID: %d, doesn't exist"
                                 .formatted(follower.getId(), followed.getId()))
                 );
         followRepository.delete(follow);
+
+        log.info("User with ID={} unfollowed user with ID={}",  follower.getId(), followed.getId());
     }
 }
